Migrate to JavaScript modules
