+++ .github/workflows/deploy-pages.yml
@@ -2,7 +2,7 @@ name: Deploy to Pages
 
 on:
   push:
-   branches: ["main"]
+    branches: ["main"]
   workflow_dispatch:
 
 permissions:
@@ -20,12 +20,23 @@ jobs:
     steps:
       - name: Checkout repository
         uses: actions/checkout@v4
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v6
+        with:
+          node-version: '24'
+
+      - name: Install dependencies
+        run: npm ci
+        
       - name: Build project
         run: make build
+
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v3
         with:
           path: ./build
+
   deploy:
     environment:
       name: github-pages
