+++ bot/game.js
@@ -1,4 +1,4 @@
-class Game {
+export class Game {
     constructor() {
         this.state = new GameState();
     }
